Configure and install Endlessh.
What is Endlessh?
Endlessh is an SSH tarpit that very slowly sends an endless, random SSH banner.
It keeps SSH clients locked up for hours or even days at a time.
The purpose is to put your real SSH server on another port and
then let the script kiddies get stuck in this tarpit instead of bothering a real server.