+ Adds moserial
Moserial is a serial terminal used for interacting with serial consoles and embedded devices.
It supports the view of incoming and outgoing data in both ASCII and
hexadecimal formats with the option to log this data.
It also supports sending and receiving of x, y, and z-modem files,
and has profiles to easily switch between configurations for different devices you are communicating with.